Characters

Description

Documentation

is_MoveInbound :: Model -> Position -> Bool

Check if the position is inside the model/environment boundaries

get_position :: Character -> MovementVector -> Position

Get the next position of a Character based on its current position, its speed & the next movement (MovementVector)

update_state :: Character -> Character

Update the state of the Character based on its energy. If the Energy is 0 or less, the state goes from Alive to Dead

update_energy :: Character -> Float -> Character

A Helper function to update the energy of a Character by a given float and make sure its stays in the bound [0,100]

get_randomMovement :: Character -> IO MovementVector

A function taking a Character and returning a potential random movement. The Character has a high chance of continuating in the same direction.

move_characterRandomly :: Model -> Character -> IO Character

Move the Character Randomely in a valid position based on its speed

update_character :: Model -> Character -> IO Character

Update the Character. This is run once every tick of the simulation