Main

Description

Documentation

background :: Color

window_w :: Int

main :: IO ()

The Main function is the opening function of our program It consists of defining the initial state of our model and starting the simulation function

argsParser :: [String] -> String -> (Bool, Int)

This function parses the arguments of the executabkes

printHelp :: IO ()

This function print the Help in the terminal