Render

Description

Tool Functions

color_dark :: Color

color_blue :: Color

color_green :: Color

color_orange :: Color

color_red :: Color

get_color :: Team -> Color

Get Color to render the Character based on its Team

float_round :: Float -> Int -> Float

Round a specific number of digit

Render Model

render_model :: Model -> IO Picture

Rendering the model by fetching the Picture of each element (Environment & [Character]) by calling their respective render function

render_UI :: Model -> Picture

Render the UI with the Picture rendered from the Tick number and the Generation Number

render_tick :: Model -> Picture

Render Tick Counter Text

render_gen :: Model -> Picture

Render Generation Counter Text

render_stats :: Model -> Picture

Render Character

render_character :: Character -> Picture

Rendering the Character by fetching the picture of its different component (Circle, Name)

render_characterCircle :: Character -> Picture

Render the Character Circle based on its position & Color

render_characterLabel :: Character -> Picture

Render the Character Name

Render Environment

render_env :: Environment -> Picture

Render the environment by fetching the pictures of each of its element (Box, Objects)

render_envBox :: Environment -> Picture

Render the box delimiting the play environment

render_food :: Food -> Picture

Render the food objects