Types

Description

Low Level Types

data Team

Constructors

T1

T2

T3

type Position = (Float, Float)

type MovementVector = (Float, Float)

data State

Constructors

Dead

Alive

type Size = Float

Model

data Model

Model contains an environment and a list of Character

Constructors

Model

Fields

environment :: Environment

character_list :: [Character]

current_tick :: Int

gen :: Int

tick_perGen :: Int

Character

data Character

Character contains a number of characteristics about its state, position and identity

Constructors

Character

Fields

name :: String

team :: Team

generation :: Int

state :: State

energy :: Float

size :: Float

speed :: Float

position :: Position

direction :: MovementVector

visionfield :: Float

basic_character :: Character

sort_character_energy :: Character -> Character -> Ordering

Environment

data Environment

Environment contains an environment size and a list of objects

Constructors

Environment

Fields

envsize :: Size

objects :: Objects

data Food

Constructors

Food

Fields

fname :: String

fsize :: Float

fenergy :: Float

fposition :: Position

basic_food :: Food

type Objects = [Food]

Objects contains all the necessary elements of the environment. For now only Food