Characters

Description

Movement Tool Functions

is_MoveInbound :: Model -> Position -> Bool

Check if the position is inside the model/environment boundaries

get_position :: Character -> MovementVector -> Position

Get the next position of a Character based on its current position, its speed & the next movement (MovementVector)

get_normalizedmovement :: MovementVector -> MovementVector

get_distance :: Position -> Position -> Float

Provide the Euclidian Distance between 2 Position

get_directionFood :: [Food] -> Character -> MovementVector

Given a list of Food Item and a character, return the direction to the closest food item

Update Fields Functions

update_state :: Character -> Character

Update the state of the Character based on its energy. If the Energy is 0 or less, the state goes from Alive to Dead

update_energy :: Character -> Float -> Character

A Helper function to update the energy of a Character by a given float and make sure its stays in the bound [0,100]

Movement Functions

get_randomMovement :: Character -> IO MovementVector

A function taking a Character and returning a potential random movement. The Character has a high chance of continuating in the same direction.

move_characterRandomly :: Model -> Character -> IO Character

Move the Character Randomely in a valid position based on its speed

move_inDirection :: MovementVector -> Character -> Character

Move the Character based on the direction input

move_character :: Model -> Character -> IO Character

Function called from the update character, executing a movement based on the environment & luck

Behavioral Functions

eat_food :: [Food] -> Character -> Character

Function updating the energy of a character based on the list of food he is eating

Character Update

update_characterAtTick :: Model -> Character -> StateT Environment IO Character

Update the Character during a normal tick of the simulation (not Tick 0 of each generation).

initR_character :: Float -> IO Character

Initiate a Character Totally Randomely

init_characterGen :: Model -> [Character] -> IO [Character]

Create a new generation based on the precedent, thanks to Genetic Algorithms

get_randomMutation :: Float -> Character -> Character -> IO Character

Create a new Character based on his parents and random mutation

Helper Functions

tool_randomtoPos :: Float -> Int -> Float -> Position

A small helper function helping to generate a position on the env box based on the boundaries, the side and a value

is_selectedGA :: Character -> IO Bool

Return if a Character is selected for Reproduction

get_pairs :: [a] -> [(a, a)]

Makes a pair out of a list. If there is an odd number of elements, the last pair is made from a tuple of the same element